# Improve performance of `cupyx.scipy.sparse.linalg.eigsh`