@anaruse 2020-10-29
Attaching the measured results.
This PR improves performance by about 25-40%.

size | nnz | nnz per row | k | scipy.eigsh | cupyx.eigsh (master branch) | cupyx.eigsh (this PR)
-----|-----|-------------|---|-------------|--------------|--------
5000 | 262483 | 52.5 | 3 | 0.101 sec | 0.108 sec | 0.087 sec
5000 | 262483 | 52.5 | 6 | 0.162 sec | 0.236 sec | 0.174 sec
5000 | 262483 | 52.5 | 12 | 0.206 sec | 0.557 sec | 0.416 sec
5000 | 477144 | 95.4 | 3 | 0.198 sec | 0.131 sec | 0.096 sec
5000 | 477144 | 95.4 | 6 | 0.252 sec | 0.235 sec | 0.185 sec
5000 | 477144 | 95.4 | 12 | 0.325 sec | 0.554 sec | 0.403 sec
5000 | 966530 | 193.3 | 3 | 0.666 sec | 0.266 sec | 0.200 sec
5000 | 966530 | 193.3 | 6 | 0.599 sec | 0.225 sec | 0.164 sec
5000 | 966530 | 193.3 | 12 | 0.761 sec | 0.611 sec | 0.447 sec
10000 | 485917 | 48.6 | 3 | 0.195 sec | 0.124 sec | 0.090 sec
10000 | 485917 | 48.6 | 6 | 0.301 sec | 0.211 sec | 0.157 sec
10000 | 485917 | 48.6 | 12 | 0.439 sec | 0.538 sec | 0.382 sec
10000 | 981164 | 98.1 | 3 | 0.519 sec | 0.205 sec | 0.147 sec
10000 | 981164 | 98.1 | 6 | 0.609 sec | 0.261 sec | 0.187 sec
10000 | 981164 | 98.1 | 12 | 0.849 sec | 0.583 sec | 0.408 sec
10000 | 1977342 | 197.7 | 3 | 1.342 sec | 0.279 sec | 0.199 sec
10000 | 1977342 | 197.7 | 6 | 1.313 sec | 0.240 sec | 0.169 sec
10000 | 1977342 | 197.7 | 12 | 1.748 sec | 0.610 sec | 0.428 sec
20000 | 964431 | 48.2 | 3 | 0.301 sec | 0.084 sec | 0.061 sec
20000 | 964431 | 48.2 | 6 | 0.576 sec | 0.205 sec | 0.150 sec
20000 | 964431 | 48.2 | 12 | 1.088 sec | 0.558 sec | 0.389 sec
20000 | 1942849 | 97.1 | 3 | 0.825 sec | 0.115 sec | 0.085 sec
20000 | 1942849 | 97.1 | 6 | 1.337 sec | 0.230 sec | 0.168 sec
20000 | 1942849 | 97.1 | 12 | 2.299 sec | 0.594 sec | 0.422 sec
20000 | 3811206 | 190.6 | 3 | 2.403 sec | 0.165 sec | 0.120 sec
20000 | 3811206 | 190.6 | 6 | 4.310 sec | 0.432 sec | 0.313 sec
20000 | 3811206 | 190.6 | 12 | 5.121 sec | 0.641 sec | 0.453 sec

scripts
```python
mport math
import numpy
import scipy
import scipy.sparse as sparse
import scipy.sparse.linalg as linalg
import cupy
import cupyx
import cupyx.scipy.sparse as cp_sparse
import cupyx.scipy.sparse.linalg as cp_linalg

dtype = numpy.float64
n_repeat = 10
n_warmup = 1


def make_sp_matrix(m, target):
    max_density = 0.0001
    min_density = 0
    while True:
        density = (min_density + max_density) / 2
        a = sparse.random(m, m, density=density, format='csr', dtype=dtype)
        a = a * a.conj().T
        nnz_per_row = a.nnz / m
        if nnz_per_row < target * 0.5:
            max_density *= math.sqrt(target / nnz_per_row)
        elif nnz_per_row < target * 0.95:
            min_density = density
        elif nnz_per_row > target * 1.05:
            max_density = density
        else:
            return a


def get_time(perf):
    cpu_time = perf.cpu_times.mean()
    gpu_time = perf.gpu_times.mean()
    return max(cpu_time, gpu_time)


def run(sp_a, k):
    times = []
    perf = cupyx.time.repeat(linalg.eigsh,
                             (sp_a, k),
                             n_repeat=n_repeat, n_warmup=n_warmup)
    times.append(get_time(perf))
    perf = cupyx.time.repeat(cp_linalg.eigsh,
                             (cp_sparse.csr_matrix(sp_a), k),
                             n_repeat=n_repeat, n_warmup=n_warmup)
    times.append(get_time(perf))
    return times


def get_time(perf):
    cpu_time = perf.cpu_times.mean()
    gpu_time = perf.gpu_times.mean()
    return max(cpu_time, gpu_time)


def run(sp_a, k):
    times = []
    perf = cupyx.time.repeat(linalg.eigsh,
                             (sp_a, k),
                             n_repeat=n_repeat, n_warmup=n_warmup)
    times.append(get_time(perf))
    perf = cupyx.time.repeat(cp_linalg.eigsh,
                             (cp_sparse.csr_matrix(sp_a), k),
                             n_repeat=n_repeat, n_warmup=n_warmup)
    times.append(get_time(perf))
    return times

umpy.random.seed(123)
cupy.random.seed(123)


print('size | nnz | nnz per row | k | scipy.eigsh | cupyx.eigsh ')
print('-----|-----|-------------|---|-------------|-------------')
for m in 5000, 10000, 20000:
    for nnz_per_m in 50, 100, 200:
        a = make_sp_matrix(m, nnz_per_m)
        for k in 3, 6, 12:
            times = run(a, k)
            print('{} | {} | {:.1f} | {} | {:.3f} | {:.3f}'
                  ''.format(m, a.nnz, a.nnz / m, k, times[0], times[1]))
```