@anaruse 2020-10-30
The performance of `cupyx.scipy.sparse.linalg.eigsh` has been further improved. 
It is about 2x faster than the current master.

size | nnz | nnz per row | k | scipy.eigsh | cupyx.eigsh 
-----|-----|-------------|---|-------------|-------------
5000 | 262483 | 52.5 | 3 | 0.103 sec | 0.051 sec
5000 | 262483 | 52.5 | 6 | 0.160 sec | 0.103 sec
5000 | 262483 | 52.5 | 12 | 0.207 sec | 0.170 sec
5000 | 477144 | 95.4 | 3 | 0.198 sec | 0.058 sec
5000 | 477144 | 95.4 | 6 | 0.254 sec | 0.116 sec
5000 | 477144 | 95.4 | 12 | 0.306 sec | 0.162 sec
5000 | 966530 | 193.3 | 3 | 0.621 sec | 0.105 sec
5000 | 966530 | 193.3 | 6 | 0.626 sec | 0.101 sec
5000 | 966530 | 193.3 | 12 | 0.842 sec | 0.236 sec
10000 | 485917 | 48.6 | 3 | 0.198 sec | 0.048 sec
10000 | 485917 | 48.6 | 6 | 0.344 sec | 0.097 sec
10000 | 485917 | 48.6 | 12 | 0.479 sec | 0.178 sec
10000 | 981164 | 98.1 | 3 | 0.644 sec | 0.073 sec
10000 | 981164 | 98.1 | 6 | 0.741 sec | 0.113 sec
10000 | 981164 | 98.1 | 12 | 0.950 sec | 0.209 sec
10000 | 1977342 | 197.7 | 3 | 1.365 sec | 0.099 sec
10000 | 1977342 | 197.7 | 6 | 1.304 sec | 0.104 sec
10000 | 1977342 | 197.7 | 12 | 1.873 sec | 0.222 sec
20000 | 964431 | 48.2 | 3 | 0.307 sec | 0.041 sec
20000 | 964431 | 48.2 | 6 | 0.612 sec | 0.092 sec
20000 | 964431 | 48.2 | 12 | 1.169 sec | 0.207 sec
20000 | 1942849 | 97.1 | 3 | 0.854 sec | 0.047 sec
20000 | 1942849 | 97.1 | 6 | 1.420 sec | 0.101 sec
20000 | 1942849 | 97.1 | 12 | 2.208 sec | 0.242 sec
20000 | 3811206 | 190.6 | 3 | 2.327 sec | 0.061 sec
20000 | 3811206 | 190.6 | 6 | 4.179 sec | 0.231 sec
20000 | 3811206 | 190.6 | 12 | 4.920 sec | 0.220 sec